lab-session6-1710/lambda-functions/requirements.txt
Pillow==10.0.1
numpy==1.24.3
boto3==1.28.62
